#8 categories - Presents valid categories to user and allows submission of a valid category to form a report

See merge request c2012125/flood-wrack-data!6

+++ src/main/java/cf/ac/uk/wrackreport/data/jpa/entities/ReportEntity.java
@@ -22,7 +22,7 @@ public class ReportEntity {
     private Long userId;
 
     @Column(name = "category_id")
-    private Long categoryId;
+    private short categoryId;
 
     @Column(name = "description")
     private String description;
